+++ common/src/main/java/cn/breadnicecat/candycraftce/block/blocks/AcidMintFlowerBlock.java
@@ -25,8 +25,8 @@ public class AcidMintFlowerBlock extends CandyPlantBlock {
 	@Override
 	public void entityInside(BlockState state, Level level, BlockPos pos, Entity entity) {
 		if (entity instanceof LivingEntity living && level.random.nextFloat() < 0.1 / TickUtils.TICK_PER_SEC) {
-			living.addEffect(new MobEffectInstance(MobEffects.POISON, 2 * TickUtils.SEC2TICK));
-			living.addEffect(new MobEffectInstance(MobEffects.CONFUSION, 2 * TickUtils.SEC2TICK));
+			living.addEffect(new MobEffectInstance(MobEffects.POISON, (int) (2 * TickUtils.SEC2TICK)));
+			living.addEffect(new MobEffectInstance(MobEffects.CONFUSION, (int) (2 * TickUtils.SEC2TICK)));
 		}
 	}
 }
