+++ forge/src/main/java/cn/breadnicecat/candycraftce/item/forge/CItemsImpl.java
@@ -18,14 +18,14 @@ import org.jetbrains.annotations.NotNull;
  */
 public class CItemsImpl {
 	public static RecordItem _recordItem(int analog, @NotNull SoundEntry evt, Item.Properties prop, int lengthInSeconds) {
-		return new RecordItem(analog, evt, prop, lengthInSeconds * TickUtils.SEC2TICK);
+		return new RecordItem(analog, evt, prop, (int) (lengthInSeconds * TickUtils.SEC2TICK));
 	}
 
 	public static RecordItem _record_wwwooowww(int analog, SoundEntry evt, Item.Properties prop, int lengthInSeconds, String nameInGame, String musicName) {
 		MutableComponent name = Component.literal(nameInGame);
 		MutableComponent music = Component.literal(musicName);
 
-		return new RecordItem(analog, evt, prop, lengthInSeconds * TickUtils.SEC2TICK) {
+		return new RecordItem(analog, evt, prop, (int) (lengthInSeconds * TickUtils.SEC2TICK)) {
 			@Override
 			public @NotNull MutableComponent getDisplayName() {
 				return music;
